+++ tonic/src/transport/service/mod.rs
@@ -2,7 +2,8 @@ mod add_origin;
 mod connection;
 mod connector;
 mod discover;
-mod grpc_timeout;
+pub(crate) mod executor;
+pub(crate) mod grpc_timeout;
 mod io;
 mod reconnect;
 mod router;
@@ -14,11 +15,11 @@ pub(crate) use self::add_origin::AddOrigin;
 pub(crate) use self::connection::Connection;
 pub(crate) use self::connector::connector;
 pub(crate) use self::discover::DynamicServiceStream;
+pub(crate) use self::executor::SharedExec;
 pub(crate) use self::grpc_timeout::GrpcTimeout;
 pub(crate) use self::io::ServerIo;
-pub(crate) use self::router::{Or, Routes};
 #[cfg(feature = "tls")]
 pub(crate) use self::tls::{TlsAcceptor, TlsConnector};
 pub(crate) use self::user_agent::UserAgent;
 
-pub use self::grpc_timeout::TimeoutExpired;
+pub use self::router::Routes;